Move circle type chart component's default datalabels config
